requirements.txt
streamlit
pandas
langchain
python-dotenv